Refactored Masthead
