+++ src/pages/index.js
@@ -18,7 +18,7 @@ import CustomButton from "../components/CustomButton"
 import FeatureBanner from "../components/FeatureBanner"
 import FeatureBannerCenter from "../components/FeatureBannerCenter"
 import Footer from "../components/Footer"
-import MastHead from "../components/MastHead"
+import MastHead from "../components/Home/MastHead"
 import MenuBar from "../components/Header/MenuBar"
 import Meta from "../components/Meta"
 import { theme } from "../styles/theme"
